backend/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-CORS==5.0.0
Flask-Limiter==4.1.1
gunicorn==25.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.2.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
